refactored node metrics
